services/api/pyproject.toml
[project]
name = "narralytica-api"
version = "0.1.0"
description = "Narralytica API service"
requires-python = ">=3.13"

dependencies = [
  # --- Web framework ---
  "fastapi>=0.128.0",
  "uvicorn[standard]>=0.40.0",

  # --- Configuration ---
  "pydantic-settings>=2.2.1",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.1",

  # --- Database ---
  "sqlalchemy>=2.0.30",
  "psycopg[binary]>=3.2.1",
  "alembic>=1.13.1",

  # --- Metrics ---
  "prometheus-client>=0.20.0",

  # --- Observability / Tracing ---
  "opentelemetry-api>=1.26.0",
  "opentelemetry-sdk>=1.26.0",
  "opentelemetry-exporter-otlp>=1.26.0",
  "opentelemetry-instrumentation-fastapi>=0.47b0",
  "opentelemetry-instrumentation-asgi>=0.47b0",
]

[tool.uv]
dev-dependencies = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
  "ruff>=0.5.0",
]
